+++ src/main/java/com/debatetimer/config/CorsConfig.java
@@ -1,8 +1,7 @@
 package com.debatetimer.config;
 
-import com.debatetimer.exception.custom.DTInitializationException;
-import com.debatetimer.exception.errorcode.InitializationErrorCode;
-import org.springframework.beans.factory.annotation.Value;
+import lombok.RequiredArgsConstructor;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
@@ -10,30 +9,16 @@ import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
+@RequiredArgsConstructor
+@EnableConfigurationProperties(CorsProperties.class)
 public class CorsConfig implements WebMvcConfigurer {
 
-    private final String[] corsOrigin;
-
-    public CorsConfig(@Value("${cors.origin}") String[] corsOrigin) {
-        validate(corsOrigin);
-        this.corsOrigin = corsOrigin;
-    }
-
-    private void validate(String[] corsOrigin) {
-        if (corsOrigin == null || corsOrigin.length == 0) {
-            throw new DTInitializationException(InitializationErrorCode.CORS_ORIGIN_EMPTY);
-        }
-        for (String origin : corsOrigin) {
-            if (origin == null || origin.isBlank()) {
-                throw new DTInitializationException(InitializationErrorCode.CORS_ORIGIN_STRING_BLANK);
-            }
-        }
-    }
+    private final CorsProperties corsProperties;
 
     @Override
     public void addCorsMappings(CorsRegistry registry) {
         registry.addMapping("/**")
-                .allowedOriginPatterns(corsOrigin)
+                .allowedOriginPatterns(corsProperties.getCorsOrigin())
                 .allowedMethods(
                         HttpMethod.GET.name(),
                         HttpMethod.POST.name(),
