+++ src/test/java/com/debatetimer/config/CorsPropertiesTest.java
@@ -9,21 +9,21 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 
-class CorsConfigTest {
+class CorsPropertiesTest {
 
     @Nested
     class Validate {
 
         @Test
         void 허용된_도메인이_null_일_경우_예외를_발생시칸다() {
-            assertThatThrownBy(() -> new CorsConfig(null))
+            assertThatThrownBy(() -> new CorsProperties(null))
                     .isInstanceOf(DTInitializationException.class)
                     .hasMessage(InitializationErrorCode.CORS_ORIGIN_EMPTY.getMessage());
         }
 
         @Test
         void 허용된_도메인이_빈_배열일_경우_예외를_발생시칸다() {
-            assertThatThrownBy(() -> new CorsConfig(new String[0]))
+            assertThatThrownBy(() -> new CorsProperties(new String[0]))
                     .isInstanceOf(DTInitializationException.class)
                     .hasMessage(InitializationErrorCode.CORS_ORIGIN_EMPTY.getMessage());
         }
@@ -31,10 +31,9 @@ class CorsConfigTest {
         @ParameterizedTest
         @NullAndEmptySource
         void 허용된_도메인_중에_빈_값이_있을_경우_예외를_발생시킨다(String empty) {
-            assertThatThrownBy(() -> new CorsConfig(new String[]{empty}))
+            assertThatThrownBy(() -> new CorsProperties(new String[]{empty}))
                     .isInstanceOf(DTInitializationException.class)
                     .hasMessage(InitializationErrorCode.CORS_ORIGIN_STRING_BLANK.getMessage());
-
         }
     }
 }
